Rename ICommandObject to CommandDTO
