+++ src/InterfaceAdapters/CommandExecutor/GameCommandDTOs/BuryCommandDTO.ts
@@ -1,0 +1,9 @@
+import CommandDTO from './CommandDTO'
+
+interface BuryCommandDTO extends CommandDTO {
+  name: 'bury'
+  params: {
+    cards: string[]
+  }
+}
+export default BuryCommandDTO
