+++ src/InterfaceAdapters/CommandExecutor/GameCommandDTOs/CommandDTO.ts
@@ -1,0 +1,6 @@
+interface CommandDTO {
+  name: string
+  params: object | null
+}
+
+export default CommandDTO
