+++ src/InterfaceAdapters/CommandExecutor/GameCommandDTOs/PlayCommandDTO.ts
@@ -1,0 +1,10 @@
+import CommandDTO from './CommandDTO'
+
+interface PlayCommandDTO extends CommandDTO {
+  name: 'play'
+  params: {
+    card: string
+  }
+}
+
+export default PlayCommandDTO
