--- src/InterfaceAdapters/CommandTypes/BuryCommandDTO.ts
@@ -1,9 +1,0 @@
-import ICommandObject from '../ICommandObject'
-
-interface BuryCommandDTO extends ICommandObject {
-  name: 'bury'
-  params: {
-    cards: string[]
-  }
-}
-export default BuryCommandDTO
