--- src/InterfaceAdapters/CommandTypes/PlayCommandDTO.ts
@@ -1,10 +1,0 @@
-import ICommandObject from '../ICommandObject'
-
-interface PlayCommandDTO extends ICommandObject {
-  name: 'play'
-  params: {
-    card: string
-  }
-}
-
-export default PlayCommandDTO
