+++ src/InterfaceAdapters/ICommandCommunicator.ts
@@ -1,9 +1,9 @@
+import CommandDTO from './CommandExecutor/GameCommandDTOs/CommandDTO'
 import ICommandCommunicatorResponse from './ICommandCommunicatorResponse'
-import ICommandObject from './ICommandObject'
 
 interface ICommandCommunicator {
-  giveCommand(command: ICommandObject): Promise<ICommandCommunicatorResponse>
   getCommands(indexOfNextCommand: number): Promise<ICommandCommunicatorResponse>
+  giveCommand(command: CommandDTO): Promise<ICommandCommunicatorResponse>
 }
 
 export default ICommandCommunicator
