--- src/InterfaceAdapters/ICommandObject.ts
@@ -1,6 +1,0 @@
-interface ICommandObject {
-  name: string
-  params: object | null
-}
-
-export default ICommandObject
